requirements.txt
python-dateutil
ariadne
uvicorn
starlette
sqlalchemy
alembic
factory_boy
aiosqlite

pytest
httpx
